+++ src/client/Microsoft.Identity.Client.Services/Platforms/net6/JsonObjectAttribute.cs
