+++ src/client/Microsoft.Identity.Client.Services/Platforms/net6/MsalJsonSerializerContext.cs
