+++ src/client/Microsoft.Identity.Client.Services/Platforms/netcore/NetCoreDeviceAuthManager.cs
