+++ src/client/Microsoft.Identity.Client.Services/Platforms/netcore/NetCoreFeatureFlags.cs
