+++ src/client/Microsoft.Identity.Client.Services/Platforms/netdesktop/NetDesktopDeviceAuthManager.cs
