+++ src/client/Microsoft.Identity.Client.Services/Platforms/netdesktop/NetDesktopFeatureFlags.cs
