+++ src/client/Microsoft.Identity.Client.Services/PlatformsCommon/Shared/ConsolePlatformLogger.cs
