+++ src/client/Microsoft.Identity.Client.Services/PlatformsCommon/Shared/DeviceAuthHelper.cs
