+++ src/client/Microsoft.Identity.Client.Services/PlatformsCommon/Shared/DeviceAuthManager.cs
