+++ src/client/Microsoft.Identity.Client.Services/PlatformsCommon/Shared/EventSourcePlatformLogger.cs
