+++ src/client/Microsoft.Identity.Client.Services/PlatformsCommon/Shared/MsalEventSource.cs
