+++ src/client/Microsoft.Identity.Client.Services/PlatformsCommon/Shared/NullDeviceAuthManager.cs
