+++ src/client/Microsoft.Identity.Client.Services/TokenResponseHelper.cs
