+++ src/client/Microsoft.Identity.Client/Extensibility/ICustomWebUI.cs
